Graceful exit on lumo compilation failure (lumo >= 1.8.0)
This patch fixes the fact that `serverless` could not display compilation
ailures because `lumo` was not returning anything. Now `lumo.build.api` accepts
a callback that returns the error.
This feature is only available in the latest lumo (>= 1.8.0) so the
README now makes clear that other versions won't be supported.