Let Serverless zip when using the lumo compiler (#29)
* Let Serverless zip when using the lumo compiler

This patch has one goal: use Serverless's packaging instead of our own for
lumo.
It has many advantages: include/exclude of files in the project
root (previously missing), auto exclusion of devDependencies and others.

Moreover, a lot of code trimming happened because of course we don't need to
use archiver anymore for zipping ourselves. As a side effect, there is no need
to save the index.js somewhere else - it will always be generated and managed
by the plugin.

* Always set default exclude and add lumo specific exclusions to it

In particular, we need to exclude .lumo_cache and the lumo binary. The latter
because of a defect in serverless that will be probably resolved at some point.

* Fix some linting issues